Add traceid to the execution time measurement
