+++ instrumentation/inslogger/logger.go
@@ -21,10 +21,15 @@ import (
 	"testing"
 
 	"github.com/insolar/insolar/core"
+	"github.com/insolar/insolar/core/utils"
 	logger "github.com/insolar/insolar/log"
 )
 
 type loggerKey struct{}
+
+func TraceID(ctx context.Context) string {
+	return utils.TraceID(ctx)
+}
 
 // FromContext returns logger from context.
 func FromContext(ctx context.Context) core.Logger {
@@ -44,7 +49,7 @@ func WithField(ctx context.Context, key string, value string) (context.Context, 
 
 // WithTraceField returns context with logger initialized with provided traceid value and logger itself.
 func WithTraceField(ctx context.Context, traceid string) (context.Context, core.Logger) {
-	ctx = setTraceID(ctx, traceid)
+	ctx = utils.SetTraceID(ctx, traceid)
 	return WithField(ctx, "traceid", traceid)
 }
 
@@ -60,21 +65,6 @@ func getLogger(ctx context.Context) core.Logger {
 		return logger.GlobalLogger
 	}
 	return l.(core.Logger)
-}
-
-type traceIDKey struct{}
-
-// TraceID returns traceid provided by WithTraceField and ContextWithTrace helpers.
-func TraceID(ctx context.Context) string {
-	val := ctx.Value(traceIDKey{})
-	if val == nil {
-		return ""
-	}
-	return val.(string)
-}
-
-func setTraceID(ctx context.Context, traceid string) context.Context {
-	return context.WithValue(ctx, traceIDKey{}, traceid)
 }
 
 // TestContext returns context with initalized log field "testname" equal t.Name() value.
