+++ logicrunner/logicrunner.go
@@ -419,8 +419,7 @@ func (lr *LogicRunner) getObjectMessage(es *ExecutionState, objref Ref) error {
 	}
 
 	var codeDesc core.CodeDescriptor
-	// TODO: [OK] pass ctx traceid
-	utils.MeasureExecutionTime("logicrunner.getObjectMessage lr.ArtifactManager.GetCode "+inslogger.TraceID(ctx), func() {
+	utils.MeasureExecutionTime(ctx, "logicrunner.getObjectMessage lr.ArtifactManager.GetCode", func() {
 		codeDesc, err = lr.ArtifactManager.GetCode(ctx, *codeRef)
 	})
 
@@ -555,8 +554,7 @@ func (lr *LogicRunner) executeConstructorCall(es *ExecutionState, m *message.Cal
 	}
 
 	var codeDesc core.CodeDescriptor
-	// TODO: [OK] use ctx trace id
-	utils.MeasureExecutionTime("logicRunner.executeConstructorCall: lr.ArtifactManager.GetCode "+inslogger.TraceID(ctx), func() {
+	utils.MeasureExecutionTime(ctx, "logicRunner.executeConstructorCall: lr.ArtifactManager.GetCode", func() {
 		codeDesc, err = lr.ArtifactManager.GetCode(ctx, *codeRef)
 	})
